Split SQL constants into different files and updated repository DDL
