+++ src/main/java/com/crunchydata/model/DataCompare.java
@@ -16,6 +16,11 @@
 
 package com.crunchydata.model;
 
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
 public class DataCompare {
     private Integer tid;
     private String tableName;
@@ -37,70 +42,6 @@ public class DataCompare {
         this.pk = pk;
         this.compareResult = compareResult;
         this.threadNbr = threadNbr;
-        this.batchNbr = batchNbr;
-    }
-
-    public Integer getTid() {
-        return tid;
-    }
-
-    public void setTid(Integer tid) {
-        this.tid = tid;
-    }
-
-    public String getTableName() {
-        return tableName;
-    }
-
-    public void setTableName(String tableName) {
-        this.tableName = tableName;
-    }
-
-    public String getPkHash() {
-        return pkHash;
-    }
-
-    public void setPkHash(String pkHash) {
-        this.pkHash = pkHash;
-    }
-
-    public String getColumnHash() {
-        return columnHash;
-    }
-
-    public void setColumnHash(String columnHash) {
-        this.columnHash = columnHash;
-    }
-
-    public String getPk() {
-        return pk;
-    }
-
-    public void setPk(String pk) {
-        this.pk = pk;
-    }
-
-    public String getCompareResult() {
-        return compareResult;
-    }
-
-    public void setCompareResult(String compareResult) {
-        this.compareResult = compareResult;
-    }
-
-    public Integer getThreadNbr() {
-        return threadNbr;
-    }
-
-    public void setThreadNbr(Integer threadNbr) {
-        this.threadNbr = threadNbr;
-    }
-
-    public Integer getBatchNbr() {
-        return batchNbr;
-    }
-
-    public void setBatchNbr(Integer batchNbr) {
         this.batchNbr = batchNbr;
     }
