+++ src/main/java/com/crunchydata/services/dbDB2.java
@@ -33,7 +33,7 @@ import java.util.Properties;
 import static com.crunchydata.services.ColumnValidation.*;
 import static com.crunchydata.util.DataUtility.ShouldQuoteString;
 import static com.crunchydata.util.DataUtility.preserveCase;
-import static com.crunchydata.util.SQLConstants.SQL_DB2_SELECT_COLUMNS;
+import static com.crunchydata.util.SQLConstantsDB2.SQL_DB2_SELECT_COLUMNS;
 import static com.crunchydata.util.Settings.Props;
 
 /**
