+++ src/main/java/com/crunchydata/services/dbMSSQL.java
@@ -32,7 +32,7 @@ import java.util.Properties;
 import static com.crunchydata.services.ColumnValidation.*;
 import static com.crunchydata.util.DataUtility.ShouldQuoteString;
 import static com.crunchydata.util.DataUtility.preserveCase;
-import static com.crunchydata.util.SQLConstants.*;
+import static com.crunchydata.util.SQLConstantsMSSQL.SQL_MSSQL_SELECT_COLUMNS;
 import static com.crunchydata.util.Settings.Props;
 
 /**
